requirements.txt
# Travel Destination Recommendation System
# Requirements for NLP + Recommendation System Project

# Core Dependencies
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.3

# NLP & Embeddings
sentence-transformers==2.3.1
transformers==4.37.2

# Machine Learning & Similarity
scikit-learn==1.4.0

# Visualization
plotly==5.18.0
folium==0.14.0
streamlit-folium==0.16.0

# AI Integration
google-generativeai==0.3.2

# Database
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.0
